+++ assets/css/admin.css
@@ -1,0 +1,212 @@
+.card {
+    border-radius: 10px;
+}
+
+.card.card-tale {
+    background: #7DA0FA;
+    color: #ffffff;
+}
+
+.card.card-tale:hover {
+    background: #96b2fb;
+}
+
+.card.card-dark-blue {
+    background: #4747A1;
+    color: #ffffff;
+}
+
+.card.card-dark-blue:hover {
+    background: #5050b2;
+}
+
+.card.card-light-blue {
+    background: #7978E9;
+    color: #ffffff;
+}
+
+.card.card-light-blue:hover {
+    background: #8f8eed;
+}
+
+.card.card-light-danger {
+    background: #F3797E;
+    color: #ffffff;
+}
+
+.card.card-light-danger:hover {
+    background: #f59095;
+}
+
+
+
+@media (min-width: 576px) {
+    .card-columns {
+        column-count: 3;
+        column-gap: 1.25rem;
+        orphans: 1;
+        widows: 1;
+    }
+
+    .card-columns .card {
+        display: inline-block;
+        width: 100%;
+    }
+}
+
+
+
+.clock {
+    width: 30rem;
+    height: 30rem;
+    border: 7px solid #545271;
+    border-radius: 50%;
+    margin: 50px auto;
+    position: relative;
+    padding: 2rem;
+    -webkit-box-shadow: 0 20px 30px rgba(104, 75, 106, 0.65);
+    -moz-box-shadow: 0 20px 30px rgba(104, 75, 106, 0.65);
+    box-shadow: 0 20px 30px rgba(104, 75, 106, 0.65);
+    background: black;
+}
+
+.outer-clock-face {
+    position: relative;
+    width: 100%;
+    height: 100%;
+    border-radius: 100%;
+    background: #fefefc;
+    -webkit-box-shadow: 0 20px 10px rgba(62, 47, 63, 0.45);
+    -moz-box-shadow: 0 20px 10px rgba(62, 47, 63, 0.45);
+    box-shadow: 0 20px 10px rgba(62, 47, 63, 0.45);
+    overflow: hidden;
+}
+
+.outer-clock-face::after {
+    -webkit-transform: rotate(90deg);
+    -moz-transform: rotate(90deg);
+    transform: rotate(90deg)
+}
+
+.outer-clock-face::before,
+.outer-clock-face::after,
+.outer-clock-face .marking {
+    content: '';
+    position: absolute;
+    width: 6px;
+    height: 100%;
+    background: #b8b8c5;
+    z-index: 0;
+    left: 49%;
+}
+
+.outer-clock-face .marking {
+    background: #bdbdcb;
+    width: 3px;
+}
+
+.outer-clock-face .marking.marking-one {
+    -webkit-transform: rotate(30deg);
+    -moz-transform: rotate(30deg);
+    transform: rotate(30deg)
+}
+
+.outer-clock-face .marking.marking-two {
+    -webkit-transform: rotate(60deg);
+    -moz-transform: rotate(60deg);
+    transform: rotate(60deg)
+}
+
+.outer-clock-face .marking.marking-three {
+    -webkit-transform: rotate(120deg);
+    -moz-transform: rotate(120deg);
+    transform: rotate(120deg)
+}
+
+.outer-clock-face .marking.marking-four {
+    -webkit-transform: rotate(150deg);
+    -moz-transform: rotate(150deg);
+    transform: rotate(150deg)
+}
+
+.inner-clock-face {
+    position: absolute;
+    top: 10%;
+    left: 10%;
+    width: 80%;
+    height: 80%;
+    background: #fefefc;
+    -webkit-border-radius: 100%;
+    -moz-border-radius: 100%;
+    border-radius: 100%;
+    z-index: 1;
+}
+
+.inner-clock-face::before {
+    content: '';
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    width: 16px;
+    height: 16px;
+    border-radius: 18px;
+    margin-left: -9px;
+    margin-top: -6px;
+    background: #4d4b63;
+    z-index: 11;
+}
+
+.hand {
+    width: 50%;
+    right: 50%;
+    height: 6px;
+    background: #61afff;
+    position: absolute;
+    top: 50%;
+    border-radius: 6px;
+    transform-origin: 100%;
+    transform: rotate(90deg);
+    transition-timing-function: cubic-bezier(0.1, 2.7, 0.58, 1);
+}
+
+.hand.hour-hand {
+    width: 20%;
+    z-index: 3;
+}
+
+.hand.min-hand {
+    height: 3px;
+    z-index: 10;
+    width: 35%;
+}
+
+.hand.second-hand {
+    background: #ff5e5e;
+    width: 40%;
+}
+
+@media screen and (max-width: 586px) {
+    .clock {
+        width: 20rem;
+        height: 20rem;
+    }
+}
+
+.center-text {
+    position: absolute;
+    top: 30%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+    font-size: 1.5em;
+    font-weight: bold;
+    text-align: center;
+}
+.center-text2 {
+    position: absolute;
+    top: 80%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+    font-size: 1.2em;
+    font-weight: bold;
+    text-align: center;
+}
